[cmake] Create static library for feature check.
