+++ src/libm-tester/check_feature.h
@@ -1,0 +1,7 @@
+#ifndef CHECK_FEATURE_H
+#define CHECK_FEATURE_H
+
+int detectFeatureSP();
+int detectFeatureDP();
+
+#endif
