+++ src/libm-tester/finite_names_testing.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include "check_feature.h"
 
 #if defined(ENABLE_SSE2) || defined(ENABLE_SSE4)
 
@@ -48,6 +49,14 @@ extern vdouble vacos_finite(vdouble);
 extern vdouble vasin_finite(vdouble);
 
 int main(void) {
+  
+  if (!detectFeatureDP() || !detectFeatureSP()) {
+    fprintf(stderr, "\n\n***** This host does not support the "
+	    "necessary CPU features to execute this program *****\n\n\n");
+    
+    return 0;
+  }
+
   double in[VECTLENDP], out[VECTLENDP];
   vdouble vin, vout;
   int i;
